+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 
  - #3693, Prevent spread embedding to allow aggregates when they are disabled - @laurenceisla
+ - #3693, A nested spread embedding now correctly groups by the fields of its top parent relationship - @laurenceisla
 
 ### Changed